Revert "Feature branch destroyer"
