+++ PAUL.yaml
@@ -2,13 +2,6 @@ maintainers:
 - Spazzy757
 # Allows for the /label command
 labels: true
-# Settings for branch destroyer
-# branch destroyer will not delete your default branch
-# set other "protected" branches here
-branch_destroyer:
-  enabled: true
-  protected_branches:
-  - main
 pull_requests:
   open_message: |
     Greetings!
