+++ pkg/config/config_test.go
@@ -1,30 +1,102 @@
 package config
 
 import (
-	"bytes"
 	"io/ioutil"
+	"os"
+	"path"
 	"testing"
-
-	"github.com/Spazzy757/paul/pkg/test"
-	"github.com/stretchr/testify/assert"
 )
 
-func TestGetPaulConfig(t *testing.T) {
-	t.Run("Test Read Paul Config Returns Valid Paul Config", func(t *testing.T) {
-		yamlFile, err := ioutil.ReadFile("../../PAUL.yaml")
-		assert.Equal(t, nil, err)
+func TestNewConfigNoSecretPath(t *testing.T) {
 
-		mClient := test.GetMockClient()
-		mClient.RepoService = &test.MockRepoService{
-			DownloadContentsResp: ioutil.NopCloser(bytes.NewReader(yamlFile)),
-		}
+	os.Setenv("SECRET_PATH", "")
 
-		owner := "test"
-		repo := "test"
-		cfg, err := GetPaulConfig(&owner, &repo, "example.com", "main", mClient)
-		assert.Equal(t, nil, err)
-		assert.NotEqual(t, "", cfg.PullRequests.OpenMessage)
-		assert.NotEqual(t, cfg.PullRequests.CatsEnabled, false)
-		assert.NotEqual(t, cfg.PullRequests.DogsEnabled, false)
-	})
+	_, err := NewConfig()
+
+	if err == nil {
+		t.Fail()
+	}
+
+	want := "SECRET_PATH env-var not set"
+	if err.Error() != want {
+		t.Errorf("want %q, got %q", want, err.Error())
+		t.Fail()
+	}
+}
+
+func TestNewConfigValidSecretPathWithApplicationID(t *testing.T) {
+	privateWant := "private"
+	secretWant := "secret"
+	appIDWant := "321"
+	tmpDir := os.TempDir()
+
+	ioutil.WriteFile(path.Join(tmpDir, "paul-private-key"), []byte(privateWant), 0600)
+	ioutil.WriteFile(path.Join(tmpDir, "paul-secret-key"), []byte(secretWant), 0600)
+
+	defer os.RemoveAll(path.Join(tmpDir, "paul-private-key"))
+	defer os.RemoveAll(path.Join(tmpDir, "paul-secret-key"))
+
+	os.Setenv("SECRET_PATH", tmpDir)
+	os.Setenv("APPLICATION_ID", appIDWant)
+
+	cfg, err := NewConfig()
+
+	if err != nil {
+		t.Errorf("%s", err.Error())
+		t.Fail()
+		return
+	}
+
+	if cfg.SecretKey != secretWant {
+		t.Errorf("want %q, got %q", secretWant, cfg.SecretKey)
+		t.Fail()
+	}
+
+	if cfg.PrivateKey != privateWant {
+		t.Errorf("want %q, got %q", privateWant, cfg.PrivateKey)
+		t.Fail()
+	}
+
+	if cfg.ApplicationID != appIDWant {
+		t.Errorf("want %q, got %q", appIDWant, cfg.ApplicationID)
+		t.Fail()
+	}
+}
+
+func TestGetFirstLine(t *testing.T) {
+	var exampleSecrets = []struct {
+		secret       string
+		expectedByte string
+	}{
+		{
+			secret:       "New-line \n",
+			expectedByte: "New-line ",
+		},
+		{
+			secret: `Newline and text 
+			`,
+			expectedByte: "Newline and text ",
+		},
+		{
+			secret:       `Example secret2 `,
+			expectedByte: `Example secret2 `,
+		},
+		{
+			secret:       "\n",
+			expectedByte: "",
+		},
+		{
+			secret:       "",
+			expectedByte: "",
+		},
+	}
+	for _, test := range exampleSecrets {
+
+		t.Run(string(test.secret), func(t *testing.T) {
+			stringNoLines := getFirstLine([]byte(test.secret))
+			if test.expectedByte != string(stringNoLines) {
+				t.Errorf("String after removal - wanted: \"%s\", got \"%s\"", test.expectedByte, test.secret)
+			}
+		})
+	}
 }
