+++ pkg/github/webhooks.go
@@ -1,33 +1,35 @@
 package github
 
 import (
+	"log"
 	"net/http"
 
-	paulclient "github.com/Spazzy757/paul/pkg/client"
 	"github.com/Spazzy757/paul/pkg/helpers"
 	"github.com/google/go-github/v32/github"
 )
 
 //IncomingWebhook handles an incoming webhook request
-func IncomingWebhook(r *http.Request, client *paulclient.GithubClient) error {
+func IncomingWebhook(r *http.Request) error {
 	// handle authentication
 	secret_key := helpers.GetEnv("SECRET_KEY", "")
 	payload, validationErr := github.ValidatePayload(r, []byte(secret_key))
 	if validationErr != nil {
 		return validationErr
 	}
-	event, parseErr := github.ParseWebHook(github.WebHookType(r), payload)
-	if parseErr != nil {
-		return parseErr
+	event, err := github.ParseWebHook(github.WebHookType(r), payload)
+	if err != nil {
+		log.Printf("could not parse webhook: err=%s\n", err)
+		return err
 	}
-	var err error
+
 	switch e := event.(type) {
 	case *github.IssueCommentEvent:
-		err = IssueCommentHandler(e, client)
+		IssueCommentHandler(e)
 	case *github.PullRequestEvent:
-		err = PullRequestHandler(e, client)
+		PullRequestHandler(e)
 	default:
-		break
+		log.Printf("unknown event type %s\n", github.WebHookType(r))
+		return nil
 	}
-	return err
+	return nil
 }
