+++ pkg/types/types_test.go
@@ -1,7 +1,11 @@
 package types
 
 import (
+	"bytes"
 	"io/ioutil"
+	"log"
+	"os"
+	"os/exec"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -11,10 +15,11 @@ func TestLoadConfig(t *testing.T) {
 	var paulConfig PaulConfig
 
 	yamlFile, err := ioutil.ReadFile("../../PAUL.yaml")
-	assert.Equal(t, nil, err)
+	if err != nil {
+		t.Errorf("yamlFile.Get err   #%v ", err)
+	}
 
-	err = paulConfig.LoadConfig(yamlFile)
-	assert.Equal(t, nil, err)
+	paulConfig.LoadConfig(yamlFile)
 	t.Run("Test Loading Config - OpenMessage", func(t *testing.T) {
 		assert.NotEqual(t, paulConfig.PullRequests.OpenMessage, "")
 	})
@@ -27,16 +32,24 @@ func TestLoadConfig(t *testing.T) {
 	t.Run("Test Loading Config - Labels", func(t *testing.T) {
 		assert.NotEqual(t, paulConfig.Labels, false)
 	})
-	t.Run("Test Loading Config - Branch Destroyer", func(t *testing.T) {
-		assert.Equal(t, true, paulConfig.BranchDestroyer.Enabled)
-		assert.Equal(t, []string{"main"}, paulConfig.BranchDestroyer.ProtectedBranches)
-	})
 
 }
 
 func TestLoadConfigFails(t *testing.T) {
 	// Only run actual test in subprocess
-	var paulConfig PaulConfig
-	err := paulConfig.LoadConfig([]byte(`% ^ & HHH`))
-	assert.NotEqual(t, nil, err)
+	if os.Getenv("SUB_PROCESS") == "1" {
+		var paulConfig PaulConfig
+		var buf bytes.Buffer
+		log.SetOutput(&buf)
+		paulConfig.LoadConfig([]byte(`% ^ & HHH`))
+		return
+	}
+	cmd := exec.Command(os.Args[0], "-test.run=TestLoadConfigFails")
+	cmd.Env = append(os.Environ(), "SUB_PROCESS=1")
+	err := cmd.Run()
+	// Cast the error as *exec.ExitError and compare the result
+	e, ok := err.(*exec.ExitError)
+	expectedErrorString := "exit status 1"
+	assert.Equal(t, true, ok)
+	assert.Equal(t, expectedErrorString, e.Error())
 }
